+++ src/main/java/org/hdf5javalib/redo/hdffile/dataobjects/messages/ObjectModificationTimeMessage.java
@@ -78,7 +78,7 @@ public class ObjectModificationTimeMessage extends HdfMessage {
         long secondsAfterEpoch = Integer.toUnsignedLong(buffer.getInt());
 
         // Return a constructed instance of ObjectModificationTimeMessage
-        return new ObjectModificationTimeMessage(version, secondsAfterEpoch, flags, (short) data.length);
+        return new ObjectModificationTimeMessage(version, secondsAfterEpoch, flags, data.length);
     }
 
     /**
